update-e-voucher

+++ src/main/java/com/example/evoucherproject/service/Impl/PurchaseServiceImpl.java
@@ -1,6 +1,9 @@
 package com.example.evoucherproject.service.Impl;
 
 import com.example.evoucherproject.exception.CustomException;
+import com.example.evoucherproject.model.dto.CustomResponse;
+import com.example.evoucherproject.model.entity.Customer;
+import com.example.evoucherproject.model.entity.Product;
 import com.example.evoucherproject.model.entity.Purchase;
 import com.example.evoucherproject.model.entity.Voucher;
 import com.example.evoucherproject.repository.CustomerRepository;
@@ -14,11 +17,9 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Service;
 
-import java.time.DayOfWeek;
-import java.time.LocalDate;
-import java.time.LocalDateTime;
-import java.time.LocalTime;
-import java.util.*;
+import java.util.ArrayList;
+import java.util.List;
+import java.util.Optional;
 
 @Service
 public class PurchaseServiceImpl implements PurchaseService {
@@ -38,11 +39,16 @@ public class PurchaseServiceImpl implements PurchaseService {
     private VoucherService voucherService;
 
     @Override
-    public List<String> paymentCustomer(int customerId, int productId) {
-        deleteExpiredVouchersByCustomerId(customerId);
-        addOrUpdateCustomerBuyProduct(customerId, productId);
-        Voucher voucher = insertOrUpdateVoucher(customerId, productId);
-        return getInformationCustomerBuyProduct(customerId, productId, voucher);
+    public CustomResponse paymentCustomer(int customerId, int productId) {
+        try {
+            deleteExpiredVouchersByCustomerId(customerId);
+            Purchase purchase = addOrUpdateCustomerBuyProduct(customerId, productId);
+            Voucher voucher = insertOrUpdateVoucher(customerId, productId, purchase);
+            return new CustomResponse("Payment Successfully!!!", HttpStatus.OK.value(), getInformationCustomerBuyProduct(customerId, productId, voucher));
+        } catch (CustomException e) {
+            return new CustomResponse(e.getMessage(), e.getHttpStatus().value(), "");
+        }
+
     }
 
     private List<String> getInformationCustomerBuyProduct(int customerId, int productId, Voucher voucher) {
@@ -88,38 +94,40 @@ public class PurchaseServiceImpl implements PurchaseService {
             }
         }
     */
-    private void addOrUpdateCustomerBuyProduct(int customerId, int productId) {
+    private Purchase addOrUpdateCustomerBuyProduct(int customerId, int productId) {
+        Optional<Purchase> purchase = purchaseRepository.getByCustomerAndProduct(customerId, productId);
+        Optional<Customer> customer = customerRepository.findById(customerId);
+        Optional<Product> product = productRepository.findById(productId);
+        if (!customer.isPresent() && !product.isPresent() && !purchase.isPresent())
+            throw new CustomException("id customer and product does not exits!", HttpStatus.NOT_FOUND);
         // nếu lần đầu mua hàng
         if (!purchaseRepository.existsByCustomerIdAndProductId(customerId, productId)) {
-            purchaseRepository.save(Purchase.builder()
+            return purchaseRepository.save(Purchase.builder()
                     .purchaseTime(DateUtils.currentDateTime())
-                    .customer(customerRepository.findById(customerId).get())
-                    .product(productRepository.findById(productId).get())
+                    .customer(customer.get())
+                    .product(product.get())
                     .quantity(1).build());
         } else {
             // cập nhật số lượng khi mua hàng
-            Optional<Purchase> purchase = purchaseRepository.getByCustomerAndProduct(customerId, productId);
-            if (!purchase.isPresent())
-                throw new CustomException("id customer and product does not exits!", HttpStatus.NOT_FOUND);
             purchase.get().setQuantity(purchase.get().getQuantity() + 1);
-            purchaseRepository.save(purchase.get());
+            return purchaseRepository.save(purchase.get());
         }
 
     }
 
-    private Voucher insertOrUpdateVoucher(int customerId, int productId) {
+    private Voucher insertOrUpdateVoucher(int customerId, int productId, Purchase purchase) {
         Optional<Voucher> voucherOptional = voucherRepository.findByCustomerCustomerId(customerId);
         // nếu voucher tồn tại  và trạng thái bằng true và  còn hàng sử dụng thì sẽ update voucher
-
         if (voucherOptional.isPresent()) {
             Voucher voucher = voucherOptional.get();
-            if (voucher.isStatus() == true && DateUtils.isDateToday(voucher.getEndTime())) {
+            // nếu tồn tại voucher và thay đổi discount nếu mua quá 5 lần
+            if (voucher.isStatus() == true && DateUtils.isDateToday(voucher.getEndTime()) || purchaseRepository.isCustomerExceededPurchaseLimit(customerId, productId) && purchase.getQuantity() < 6) {
                 voucher.setDiscount(7);
             }
             return voucherRepository.save(voucher);
 
         } else {
-            if (purchaseRepository.isCustomerExceededPurchaseLimit(customerId, productId)) {
+            if (purchaseRepository.isCustomerExceededPurchaseLimit(customerId, productId) && purchase.getQuantity() < 6) {
                 // còn nếu chưa tồn tại mà khách hàng mua 5 lần thì nhận 1 voucher mới
                 voucherService.saveVoucher(customerId, 7);
             }
@@ -129,10 +137,8 @@ public class PurchaseServiceImpl implements PurchaseService {
 
     private void deleteExpiredVouchersByCustomerId(int customerId) {
         Optional<Voucher> isVoucher = voucherRepository.findByCustomerCustomerId(customerId);
-        if (isVoucher.isPresent()) {
-            if (DateUtils.isDateToday(isVoucher.get().getEndTime())) {
-                voucherRepository.delete(isVoucher.get());
-            }
+        if (isVoucher.isPresent() && DateUtils.isDateToday(isVoucher.get().getEndTime())) {
+            voucherRepository.delete(isVoucher.get());
         }
     }
 }
